Change to polygon-based drawing and implement basis for snowflake
Call GL_TRIANGLES instead of GL_POINTS when drawing. Utilize a recursive method to calculate the vertices of the triangles.

For initial work, bypass the processTriangle method and directly call the drawTriangle method. This draws a simple white triangle on the screen. The recursive processTriangle method will need to be modified to accommodate the Koch snowflake calculations.