Implement processing and drawing of Koch snowflake
Create `koch` method, which takes three initial points and calls the new processLine method three times.

Create `processLine` method, which takes two points and a recursion counter. It calculates the points necessary for the current level of recursion and calls itself four times, since there are four lines to be processed between five points (the original two points and the three newly calculated ones).

Create `drawLine` method, which takes two points and passes them to `GL_LINES`.

Remove `distanceBetween` method (no longer needed).

Change `sideLength` to a member variable.

Change `m_vertexPositions`'s dimensions to 4 instead of 6, since we are drawing lines, not triangles.

Add code to `main` to accept user input for the level of recursion and side length of the original triangle.

Create `displayCalculations` method, which takes a level of recursion (an int) and displays it, the number of sides of the triangle, the perimeter, and the area at the specified recursion level.

Call the `displayCalculations` method repeatedly at the beginning of the `Project1` constructor to display statistics for each level of recursion up to the value input by the user.